Major Update Towards PapaTebs
Within the code, there was a major bug which caused mobile phones and tablets to display the site with text collided with each other. A temporary fix was created to not display the 'hamburger button' logo.

Now, the hamburger button was reimplemented and the code is now adjusted and organized to have less code.